backend/requirements.txt
# Core Flask backend
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
requests==2.31.0

# PDF generation
reportlab==4.0.7
pillow==10.1.0

# 🔥 REVOLUTIONARY DOCUMENT PARSING
PyPDF2==3.0.1
pdfplumber==0.10.3
pytesseract==0.3.10
opencv-python==4.8.1.78
python-dateutil==2.8.2

# 🧠 spaCy NLP for medical entity recognition (fast, offline, accurate!)
# Works on Grandma Anne's potato MacBook!
spacy>=3.7.0
# Download model after install: python -m spacy download en_core_web_sm

# Data processing (minimal set)
numpy>=1.26.0
pandas>=2.2.0
matplotlib>=3.8.0

# API clients (keep for future flexibility)
openai==1.3.7
